will no longer run velvet assembly if the longest contig output is found in the directory

+++ initialVelvet.sh
@@ -5,12 +5,17 @@
 
 # run velvet on the reads that mapped to COI and pull out the longest contig that was assembled
 # this longest contig will be the input to aTRAM
+# getLongest.py needs to be in the path
 
 for f in *COI.fq;
 do
-velveth $f.initial 29 -short -fastq $f;
-velvetg $f.initial;
-python2.7 getLongest.py -f $f.initial/contigs.fa -o $f.initial.longest.fasta;
-rm -r $f.initial;
+	if [  -e $f.initial.longest.fasta  ];
+		then echo Initial assembly of $f reads already completed.;
+	else
+		velveth $f.initial 29 -short -fastq $f;
+		velvetg $f.initial;
+		python2.7 getLongest.py -f $f.initial/contigs.fa -o $f.initial.longest.fasta;
+		rm -r $f.initial;
+	fi;
 done
